Add test for python 2.7.

+++ ci/do_ci.sh
@@ -153,11 +153,22 @@ elif [[ "$1" == "plugin.test" ]]; then
   [[ "$NUM_SPANS" -eq "1" ]] || exit 1
   exit 0
 elif [[ "$1" == "python.wheel.test" ]]; then
+  # Python 3
   python3 -m pip install /plugin/*cp3*.whl
   MOCK_SATELLITE_PORT=5000
   /plugin/mock_satellite $MOCK_SATELLITE_PORT &
   MOCK_SATELLITE_PID=$!
   python3 test/bridge/python/span_probe.py 127.0.0.1 $MOCK_SATELLITE_PORT
+  NUM_SPANS=`/plugin/mock_satellite_query --port $MOCK_SATELLITE_PORT num_spans`
+  kill $MOCK_SATELLITE_PID
+  [[ "$NUM_SPANS" -eq "1" ]] || exit 1
+
+  # Python 2.7
+  python -m pip install /plugin/*cp27mu*.whl
+  MOCK_SATELLITE_PORT=5000
+  /plugin/mock_satellite $MOCK_SATELLITE_PORT &
+  MOCK_SATELLITE_PID=$!
+  python test/bridge/python/span_probe.py 127.0.0.1 $MOCK_SATELLITE_PORT
   NUM_SPANS=`/plugin/mock_satellite_query --port $MOCK_SATELLITE_PORT num_spans`
   kill $MOCK_SATELLITE_PID
   [[ "$NUM_SPANS" -eq "1" ]] || exit 1
